requirements.txt
# Jira Agent - Requirements (New Architecture)
# Google ADK and AI dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0

# Jira integration
jira>=3.5.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utility libraries
python-dotenv>=1.0.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Code quality (development)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Documentation (development)
mkdocs>=1.5.0
mkdocs-material>=9.4.0